requirements.txt
requests
python-dotenv
openai
gradio
